refactor(topic-right-pane): lazy load branch flow panel
