+++ src/renderer/pages/home/components/TopicRightPane.tsx
@@ -1,4 +1,5 @@
-import type { TopicMessageFlowLiveState } from '@renderer/components/chat/flow'
+// eslint-disable-next-line barrel/closed -- Bypass the flow barrel so chat startup does not touch TopicMessageFlowCanvas.
+import type { TopicMessageFlowLiveState } from '@renderer/components/chat/flow/topicMessageFlowLiveTree'
 import {
   RESOURCE_PANE_TAB,
   type ResourcePaneConfig,
@@ -17,10 +18,10 @@ import { useIsActiveTab } from '@renderer/hooks/tab'
 import { useWindowFrame } from '@renderer/hooks/useWindowFrame'
 import { Activity, GitBranch } from 'lucide-react'
 import type { PropsWithChildren } from 'react'
-import { createContext, use, useCallback, useRef, useSyncExternalStore } from 'react'
+import { createContext, lazy, Suspense, use, useCallback, useRef, useSyncExternalStore } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import TopicBranchPanel from './TopicBranchPanel'
+const TopicBranchPanel = lazy(() => import('./TopicBranchPanel'))
 
 interface TopicRightPaneSurfaceProps {
   topicId?: string
@@ -179,17 +180,19 @@ function TopicRightPaneSurface({
       <ResourcePanePanel />
       {hasBranchPanel && (
         <Shell.Panel value="branch">
-          <TopicBranchPanel
-            open
-            topicId={topicId}
-            topicName={topicName}
-            liveState={branchLiveState}
-            focusKey={canvasFocusKey}
-            layoutReady={canvasLayoutReady}
-            onLocateMessage={handleLocateMessage}
-            onStartBranchDraft={onStartBranchDraft}
-            onCancelBranchDraft={onCancelBranchDraft}
-          />
+          <Suspense fallback={null}>
+            <TopicBranchPanel
+              open
+              topicId={topicId}
+              topicName={topicName}
+              liveState={branchLiveState}
+              focusKey={canvasFocusKey}
+              layoutReady={canvasLayoutReady}
+              onLocateMessage={handleLocateMessage}
+              onStartBranchDraft={onStartBranchDraft}
+              onCancelBranchDraft={onCancelBranchDraft}
+            />
+          </Suspense>
         </Shell.Panel>
       )}
       {hasBranchPanel && enableDeveloperMode && (
